+++ frontend/src/components/UI/Squares.tsx
@@ -1,15 +1,21 @@
+import React from "react";
 import { createPortal } from "react-dom";
-import styles from "./Squares.module.css"
+import styles from "./Squares.module.css";
 
-const Squares = () => {
-    return createPortal(<div className={styles["square-container"]}>
-    <div className={styles.square}></div>
-    <div className={styles.square}></div>
-    <div className={styles.square}></div>
-    <div className={styles.square}></div>
-    <div className={styles.square}></div>
-    <div className={styles.square}></div>
-</div>, document.getElementById("squares") as HTMLElement);
-}
+const Squares: React.FC<React.PropsWithChildren> = (
+	props: React.PropsWithChildren
+) => {
+	return (
+		<div className={styles["square-container"]}>
+			<div className={styles.square}></div>
+			<div className={styles.square}></div>
+			<div className={styles.square}></div>
+			<div className={styles.square}></div>
+			<div className={styles.square}></div>
+			<div className={styles.square}></div>
+			{props.children}
+		</div>
+	);
+};
 
 export default Squares;