+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat(components-login): Login page added
 * 🎉 feat(components-squares): Squares component added
 * 🎉 feat(login): fully responsive login page styles added